update go version

+++ grid-client/deployer/gateway_fqdn_deployer_test.go
@@ -307,7 +307,7 @@ func TestFQDNDeployer(t *testing.T) {
 		err := d.Sync(context.Background(), &gw)
 		assert.Error(t, err)
 		assert.Equal(t, gw.NodeDeploymentID, map[uint32]uint64{nodeID: contractID})
-		assert.Equal(t, gw.ContractID, uint64(contractID))
+		assert.Equal(t, gw.ContractID, contractID)
 	})
 
 	t.Run("test sync", func(t *testing.T) {
@@ -336,7 +336,7 @@ func TestFQDNDeployer(t *testing.T) {
 		err = d.Sync(context.Background(), &gw)
 		assert.NoError(t, err)
 		assert.Equal(t, gw.NodeDeploymentID, map[uint32]uint64{nodeID: contractID})
-		assert.Equal(t, gw.ContractID, uint64(contractID))
+		assert.Equal(t, gw.ContractID, contractID)
 		assert.Equal(t, gw.FQDN, "name.com")
 	})
 
@@ -364,7 +364,7 @@ func TestFQDNDeployer(t *testing.T) {
 		err = d.Sync(context.Background(), &gw)
 		assert.NoError(t, err)
 		assert.Equal(t, gw.NodeDeploymentID, map[uint32]uint64{nodeID: contractID})
-		assert.Equal(t, gw.ContractID, uint64(contractID))
+		assert.Equal(t, gw.ContractID, contractID)
 		assert.Equal(t, gw.FQDN, "")
 		assert.Equal(t, gw.Name, "")
 		assert.Equal(t, gw.TLSPassthrough, false)
